% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_colnames}
\alias{check_colnames}
\title{Check column names}
\usage{
check_colnames(avail_colnames, req_colnames)
}
\arguments{
\item{avail_colnames}{Available column names to check if include
'req_colnames'}

\item{req_colnames}{Required column names to check if are among
'avail_colnames'}
}
\description{
Check column names
}
