% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package}
\alias{check_package}
\title{Check package}
\usage{
check_package(package, repo = "CRAN", git_repo = "", required = TRUE)
}
\arguments{
\item{package}{Name of package}

\item{repo}{Repository of package, choose from "CRAN", "Bioc", and "github"}

\item{git_repo}{If repo is github, name of the GitHub repository}

\item{required}{Whether the package is required - will break the program if
required but not installed.}
}
\value{
Boolean - TRUE if package is present, FALSE if not.
}
\description{
Check package
}
