% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_markers}
\alias{example_markers}
\title{The marker panel of \code{example_reference} and \code{example_query}, i.e., the panel
of the Z2YR CyTOF dataset.}
\format{
A character vector of length 42
}
\usage{
example_markers
}
\description{
The marker panel of \code{example_reference} and \code{example_query}, i.e., the panel
of the Z2YR CyTOF dataset.
}
\keyword{datasets}
