% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model}
\alias{example_model}
\title{An example of a model saved by cyDefine}
\format{
A tibble of 161,764 rows and 244 variables
}
\usage{
example_model
}
\description{
Trained using caret::train(). For testing purposes.
}
\keyword{datasets}
