% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_query}
\alias{example_query}
\title{An example of a CyTOF query, already prepared by cyDefine as seen in
´data-raw/example_query.R´}
\format{
A tibble of 79,179 rows and 45 columns
}
\source{
\url{http://flowrepository.org/id/FR-FCM-Z2YR}
}
\usage{
example_query
}
\description{
An example of a CyTOF query, already prepared by cyDefine as seen in
´data-raw/example_query.R´
}
\keyword{datasets}
