% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_reference}
\alias{example_reference}
\title{An example of a CyTOF reference, already prepared by cyDefine as seen in
´data-raw/example_reference.R´}
\format{
A tibble of 80,405 rows and 46 columns
}
\source{
\url{http://flowrepository.org/id/FR-FCM-Z2YR}
}
\usage{
example_reference
}
\description{
An example of a CyTOF reference, already prepared by cyDefine as seen in
´data-raw/example_reference.R´
}
\keyword{datasets}
