% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{excl_redundant_populations}
\alias{excl_redundant_populations}
\title{Exclude Redundant Cell Populations from Reference}
\usage{
excl_redundant_populations(
  reference,
  query,
  markers,
  num.threads = 4,
  mtry = 22,
  min_cells = 50,
  min_pct = 0.001,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers present in both query
and reference.}

\item{num.threads}{Number of threads. Use 0 for all available cores. Default is 2 if not set by options/environment variables (see below).}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{min_cells}{An integer specifying the minimum number of cells for a cell type to be retained. Default is 50.}

\item{min_pct}{A numeric value specifying the minimum percentage of cells for a cell type to be retained. Default is 0.001.}

\item{seed}{desc}

\item{verbose}{Verbosity}
}
\value{
A data frame containing the filtered reference dataset with redundant cell types excluded.
}
\description{
This function filters out redundant cell types from a reference dataset based on their counts and proportions in the query dataset. It uses an initial projection to classify the cell types and then excludes those with counts below a specified minimum or a proportion below a specified percentage.
}
