% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{get_distinct_colors}
\alias{get_distinct_colors}
\title{Get distinct colors for a range of cell types}
\usage{
get_distinct_colors(populations, add_unassigned = TRUE)
}
\arguments{
\item{populations}{Character vector of populations to get colors for}

\item{add_unassigned}{Boolean indicating whether the color 'black' should be
added for unassigned cells}
}
\value{
Named list with values HEX colors and populations as names
}
\description{
Get distinct colors for a range of cell types
}
