% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{get_merge_list}
\alias{get_merge_list}
\title{Get Merge List for create_diagram}
\usage{
get_merge_list(adapted_reference)
}
\arguments{
\item{adapted_reference}{A \code{data.frame} or \code{tibble} that contains the adapted reference data.
The data should include the columns \code{cluster}, \code{celltype}, and \code{celltype_original}.
These columns are expected to be present in the dataset after running \code{adapt_reference()}.}
}
\value{
A named list where each element represents a merged cell type. The names of the list
correspond to the unique cell types in the \code{celltype} column, and the values are lists of
the original cell types (\code{celltype_original}) that were merged into the final cell type.
}
\description{
This function generates a list mapping merged cell types to their original cell types
based on the clustering results from an adapted reference dataset. The function assumes
that the dataset has been preprocessed using \code{adapt_reference()}.
}
\details{
The function filters the provided reference data to include only unique combinations
of \code{cluster}, \code{celltype}, and \code{celltype_original}. It then groups the data by \code{cluster} and
creates a list where each unique \code{celltype} is associated with its corresponding original
cell types. The list is returned in a format suitable for use in generating flowcharts or
diagrams that visualize the merging of cell types.
}
\examples{
\dontrun{
# Assuming `adapted_reference` is a preprocessed tibble from `adapt_reference()`
merge_list <- get_merge_list(adapted_reference)
}

}
