% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReference}
\alias{getReference}
\title{Helper function to get spaceRAT scaffolds}
\usage{
getReference(
  name = "pbmc",
  store = FALSE,
  path = "data",
  timeout = 120,
  verbose = TRUE
)
}
\arguments{
\item{name}{Name of scaffold to get}

\item{store}{Whether or not to store the scaffold for future use}

\item{path}{Path to folder for storing scaffolds}

\item{timeout}{Setting download timeout}

\item{verbose}{Verbosity}
}
\value{
A spaceRAT scaffold
}
\description{
This function acquires a given prebuilt scaffold from the scaffolds
available on Zenodo. To prevent repeated downloads, \code{store} can be set to
\code{TRUE}. This will create a "scaffolds" folder, where the scaffolds are
stored.
}
