% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_identify_unassigned.R
\name{identify_unassigned}
\alias{identify_unassigned}
\title{Identify unassigned cells}
\usage{
identify_unassigned(
  reference,
  query = NULL,
  markers,
  model = NULL,
  num.threads = 1,
  mtry = ceiling(length(markers)/3),
  identify_type = "probability",
  probability_threshold = 0.8,
  train_on_unassigned = FALSE,
  unassigned_name = "unassigned",
  pct_expl_var = 0.95,
  MAD_factor = 2.5,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of classified query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{model}{Random forest model for \verb{identify_type = "probability}.}

\item{num.threads}{Number of threads. Use 0 for all available cores. Default is 2 if not set by options/environment variables (see below).}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{identify_type}{(Default: probability) Type of outlier detection.
Options are (multiple allowed in conjunction):
"probability" - Cut cells with a lower probability than the \code{probability_threshold}.
"trained" - Train on unassigned in the reference.
"mad" - MAD threshold on mahalanobis distance in PC space per cell type.
"maha" - Same as "mad", but in marker space.
"sd" - Unassign cells with more than 20\% of markers that lay more than two standard deviations from reference expression}

\item{probability_threshold}{Threshold for random forest probability threshold. Use in conjunction with \code{unassigned_type = "probability"}.}

\item{train_on_unassigned}{Boolean indicating whether unassigned cells should be included in model training.
Recommended when reference and query are samples stemming from the same experiment and unassigned
cells of the query are assumed to be representative of those found in the reference.}

\item{unassigned_name}{Name used for unassigned cells}

\item{pct_expl_var}{The percentage threshold of explained variance for PC selection}

\item{MAD_factor}{Median average distance threshold of the outlier to its assigned population}

\item{seed}{Random seed}

\item{verbose}{Verbosity}
}
\value{
A tibble of unassigned cells
}
\description{
Identify unassigned cells
}
