% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_classify_cells.R
\name{macroF1_summary}
\alias{macroF1_summary}
\title{Summary function for using macro-average F1-score as metric in caret::trainControl}
\usage{
macroF1_summary(data, lev, model = NULL)
}
\arguments{
\item{data}{a data frame with columns \code{obs} and
\code{pred} for the observed and predicted outcomes. For metrics
that rely on class probabilities, such as
\code{twoClassSummary}, columns should also include predicted
probabilities for each class. See the \code{classProbs} argument
to \code{\link[caret]{trainControl}}.}

\item{lev}{a character vector of factors levels for the
response. In regression cases, this would be \code{NULL}.}

\item{model}{a character string for the model name (as taken
from the \code{method} argument of \code{\link[caret]{train}}.}
}
\value{
Macro-average F1 score
}
\description{
Summary function for using macro-average F1-score as metric in caret::trainControl
}
