% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_identify_unassigned.R
\name{MAD_max_distance}
\alias{MAD_max_distance}
\title{Use median absolute deviation to find the upper boundary for outlier detection}
\usage{
MAD_max_distance(distances, MAD_factor = 3)
}
\arguments{
\item{distances}{Numeric vector of Mahalanobis distances}

\item{MAD_factor}{Coefficient used to define the boundary. Generally, 3 is
considered very conservative, 2.5 is moderately conservative, and 2 is poorly
conservative.}
}
\value{
Upper limit
}
\description{
Use median absolute deviation to find the upper boundary for outlier detection
}
