% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_identify_unassigned.R
\name{mahalanobis_distances}
\alias{mahalanobis_distances}
\title{Compute Mahalanobis distance from each cell to a given population}
\usage{
mahalanobis_distances(cells, population, weights = NULL)
}
\arguments{
\item{cells}{Tibble of cells (cells in rows, markers in columns) from each of which the Mahalanobis distance to the population should be computed. ???}

\item{population}{Tibble of a population of cells (cells in rows, markers in columns) to which the Mahalanobis distance should be computed. ???}

\item{weights}{Numeric vector of marker weights for computing the weighted Mahalanobis distance. ???}
}
\value{
Numeric vector of Mahalanobis distance to population for each cell.
}
\description{
Compute Mahalanobis distance from each cell to a given population
}
