% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{map_marker_names}
\alias{map_marker_names}
\title{Modify query marker names to match reference marker names}
\usage{
map_marker_names(
  query,
  query_markers,
  ref_markers = cyDefine::pbmc_markers,
  using_pbmc = TRUE,
  map_specific_from = NULL,
  map_specific_to = NULL,
  return_unrecognized = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{query_markers}{Character vector of query markers}

\item{ref_markers}{Character vector of reference markers}

\item{using_pbmc}{Boolean indicating whether or not the Universal PBMC atlas is
used as reference}

\item{map_specific_from}{Optional: used together with \code{map_specific_to}.
Character vector of specific marker(s) of the query to manually map to
specific marker(s) of the reference (given by \code{map_specific_to}).}

\item{map_specific_to}{Optional: used together with \code{map_specific_from}.
Markers of the reference that the markers given by \code{map_specific_from} should
be mapped to.}

\item{return_unrecognized}{Boolean specifying whether to return names of
unrecognized markers.}

\item{verbose}{Verbosity}
}
\value{
Tibble of query data modified to make marker names match reference.
If return_unrecognized is TRUE, returns a list with this tibble and a
character vector with the unrecognized markers.
}
\description{
Modify query marker names to match reference marker names
}
\examples{

map_marker_names(
  using_pbmc = TRUE, query = example_query,
  query_markers = example_markers, map_specific_from = c("CXCR5", "TCRgd"),
  map_specific_to = c("CD185", "gdTCR")
)

}
\seealso{
Other adapt: 
\code{\link{get_merged_label}()},
\code{\link{merge_populations}()}
}
\concept{adapt}
