% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{merge_populations}
\alias{merge_populations}
\title{Merge groups of similar cell populations}
\usage{
merge_populations(populations_to_merge, reference)
}
\arguments{
\item{populations_to_merge}{Tibble ???}

\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{name_unassigned_if_merged}{Cell types of the labeling tree, where if merged, the merged population should rather be named 'unassigned'}
}
\description{
Merge groups of similar cell populations
}
\seealso{
Other adapt: 
\code{\link{get_merged_label}()},
\code{\link{map_marker_names}()}
}
\concept{adapt}
