% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers}
\alias{pbmc_markers}
\title{Markers of the Seurat PBMC atlas}
\format{
A character vector of length 241
}
\usage{
pbmc_markers
}
\description{
Markers of the Seurat PBMC atlas
}
\keyword{datasets}
