% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_reference}
\alias{pbmc_reference}
\title{The modified Seurat PBMC atlas}
\format{
A tibble of 161,764 rows and 244 variables
}
\source{
\url{https://atlas.fredhutch.org/data/nygc/multimodal/pbmc_multimodal.h5seurat}
}
\usage{
pbmc_reference
}
\description{
Load the modified PBMC reference with
\code{pbmc_reference <- get_reference("pbmc")}
See \code{?get_reference} for options to store the reference locally for subsequent faster load times.
}
\keyword{datasets}
