% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{plot_alluvium}
\alias{plot_alluvium}
\title{Create an Alluvial Plot}
\usage{
plot_alluvium(
  df,
  true = "celltype",
  predicted = "model_prediction",
  color = NULL,
  color_by = c("true", "predicted"),
  n = 500,
  seed = 1286
)
}
\arguments{
\item{df}{A data frame containing the classified data.}

\item{true}{A character string indicating the column name for true labels.}

\item{predicted}{A character string indicating the column name for predicted labels.}

\item{color}{A named vector of colors for the true labels.}

\item{color_by}{A character string indicating which variable to use for filling the alluvia ("true" or "predicted").}

\item{n}{An integer indicating the sample size.}

\item{seed}{An integer indicating the seed value for random sampling.}
}
\value{
A ggplot object representing the alluvial plot.
}
\description{
This function creates an alluvial plot to compare true and predicted clusterings.
}
