% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap of marker expressions per population}
\usage{
plot_heatmap(
  data,
  population_col = "predicted_celltype",
  markers_to_plot,
  title = "Average marker expression of predicted cell types",
  return_data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Tibble with data (cells in rows, markers in columns)}

\item{population_col}{Name of the cell type column in 'data' that you want to
plot expressions for}

\item{markers_to_plot}{Character vector of the markers that you want to
include in the heatmap}

\item{title}{Title of the plot}

\item{return_data}{Boolean indicating if the data used for plotting should be
returned rather than a ggplot2 object}

\item{verbose}{Verbosity}
}
\description{
Plot heatmap of marker expressions per population
}
