% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_reference}
\alias{seurat_reference}
\title{The modified Seurat PBMC atlas}
\format{
A tibble of 161,764 rows and 244 variables
}
\source{
\url{https://atlas.fredhutch.org/data/nygc/multimodal/pbmc_multimodal.h5seurat}
}
\usage{
seurat_reference
}
\description{
The modified Seurat PBMC atlas
}
\keyword{datasets}
