% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{adapt_reference}
\alias{adapt_reference}
\title{Adapt Seurat PBMC reference to query marker panel}
\usage{
adapt_reference(
  reference,
  query,
  markers,
  using_seurat = TRUE,
  exclude_celltypes = c("Doublet", "Platelet", "Eryth", "HSPC"),
  celltype_col = ifelse(using_seurat, "celltype.l2", "celltype"),
  min_f1 = 0.7,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers present in both query
and reference.}

\item{using_seurat}{Boolean indicating whether or not the Seurat PBMC atlas is
used as reference}

\item{exclude_celltypes}{Only relevant if \code{using_seurat = TRUE}. Character
vector of cell types of the Seurat reference to NOT consider during cell
type classification. Defaults to non-PBMCs.}

\item{min_f1}{Minimum F1 score required for two cell types to be separated}

\item{verbose}{Verbosity}
}
\value{
Tibble of adapted reference
}
\description{
Adapt Seurat PBMC reference to query marker panel
}
