% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{adapt_reference}
\alias{adapt_reference}
\title{Adapt Seurat PBMC reference to query marker panel}
\usage{
adapt_reference(
  avail_markers,
  exclude_celltypes = c("Doublet", "Platelet", "Eryth", "HSPC"),
  min_f1 = 0.7,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{avail_markers}{Character vector of available markers (intersecting query and reference)}

\item{exclude_celltypes}{Character vector of cell types of the Seurat reference to NOT consider for the specific data. Defaults to non-PBMCs.}

\item{min_f1}{Minimum F1 score required for two cell types to be separated}

\item{verbose}{Verbosity}

\item{reference}{Tibble of reference data (cells in rows, markers in columns)}
}
\value{
Tibble of adapted reference
}
\description{
Adapt Seurat PBMC reference to query marker panel
}
