% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Perform batch correction via cyCombine}
\usage{
batch_correct(
  reference,
  query,
  markers,
  covar = NULL,
  norm_method = "scale",
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input tibble.
If NULL, no covar will be used}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'scale'}

\item{seed}{The seed to use when creating the SOM.}

\item{verbose}{Verbosity}
}
\value{
Named list of tibbles of batch corrected reference and query data
}
\description{
Perform batch correction via cyCombine
}
