% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Perform batch correction via cyCombine}
\usage{
batch_correct(
  reference,
  query,
  markers,
  covar = NULL,
  label = NULL,
  norm_method = "scale",
  ref.batch = NULL,
  seed = 332,
  xdim = 8,
  ydim = 8,
  rlen = 10,
  mc.cores = NULL,
  num.threads = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{covar}{Optional: Covariate column name or vector for correction.}

\item{label}{Character string for the cluster labels column. If this column
already exists in the object metadata, the normalization and clustering steps are
skipped, and correction proceeds using these existing labels. Defaults to "Labels".}

\item{norm_method}{Normalization method ("scale", "rank", "clr", "lognorm", "qnorm", "none").
Applied before clustering unless \code{labels_colname} is provided. Defaults to "scale".}

\item{ref.batch}{Optional: Reference batch for ComBat.}

\item{seed}{Integer, random seed for reproducibility.}

\item{xdim}{Integer or vector of integers, the x-dimension(s) of the SOM grid
(for "kohonen" and "flowsom"). If a vector, the workflow iterates.}

\item{ydim}{Integer or vector of integers, the y-dimension(s) of the SOM grid.
If a vector, the workflow iterates. Length should match \code{xdim} or be 1.}

\item{rlen}{Integer, the number of iterations for SOM training.}

\item{mc.cores}{Number of cores for parallel processing. Defaults to 1.}

\item{num.threads}{Number of threads to parallelize with}

\item{verbose}{Verbosity}

\item{...}{Additional optional parameters for batch correction.
See \code{?cyCombine::batch_correct} for options.}
}
\value{
Named list of tibbles of batch corrected reference and query data
}
\description{
Perform batch correction via cyCombine
}
