% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_classify_cells.R
\name{classify_cells}
\alias{classify_cells}
\title{Classify canonical cell types}
\usage{
classify_cells(
  reference,
  query,
  markers,
  subsample = 2000,
  mtry = floor(length(markers)/3),
  min.node.size = 1,
  splitrule = "gini",
  num.trees = 300,
  use.weights = FALSE,
  probability = TRUE,
  unassigned_name = "unassigned",
  load_model = NULL,
  save_model = NULL,
  num.threads = 1,
  return_pred = FALSE,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{subsample}{Number of cells to use for training from each cell type}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{min.node.size}{Minimal node size to split at. Default 1 for classification, 5 for regression, 3 for survival, and 10 for probability. For classification, this can be a vector of class-specific values.}

\item{splitrule}{Splitting rule. For classification and probability estimation "gini", "extratrees" or "hellinger" with default "gini".
For regression "variance", "extratrees", "maxstat", "beta" or "poisson" with default "variance".
For survival "logrank", "extratrees", "C" or "maxstat" with default "logrank".}

\item{num.trees}{Number of trees.}

\item{use.weights}{(Default: FALSE) Boolean whether to use celltype frequency as class weights}

\item{probability}{Grow a probability forest as in Malley et al. (2012).}

\item{unassigned_name}{Name used for unassigned cells}

\item{load_model}{Optional: Path to an rda file of a previously trained model}

\item{save_model}{Optional: Path to save an rda file of the trained model}

\item{num.threads}{Number of threads. Use 0 for all available cores. Default is 2 if not set by options/environment variables (see below).}

\item{return_pred}{Boolean indicating if only predictions should be returned as a character vector}

\item{seed}{Random seed}

\item{verbose}{Verbosity}
}
\value{
Tibble of query data with an added column of the predicted cell type, 'model_prediction'
}
\description{
Classify canonical cell types
}
