% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_classify_cells.R
\name{classify_cells}
\alias{classify_cells}
\title{Classify canonical cell types}
\usage{
classify_cells(
  reference,
  query,
  markers,
  unassigned_name = "ungated",
  load_model = NULL,
  save_model = NULL,
  return_pred = FALSE,
  n_cv_folds = 5,
  n_trees = 100,
  seed = 332,
  n_cores = 2,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{unassigned_name}{Name used for unassigned cells}

\item{load_model}{Optional: Path to an RDS file of a previously trained model}

\item{save_model}{Optional: Path to save an RDS file of the trained model}

\item{return_pred}{Boolean indicating if only predictions should be returned as a character vector}

\item{n_cv_folds}{Number of cross-validation folds used for hyperparameter tuning}

\item{n_trees}{Number of trees in the random forest}

\item{seed}{Random seed}

\item{n_cores}{The number of cores to use for parallel execution. Obs:
parallel execution only performed if the 'doParallel' package is installed!}

\item{verbose}{Verbosity}

\item{train_on_unassigned}{Boolean indicating whether unassigned cells should be included in model training.
Recommended when reference and query are samples stemming from the same experiment and unassigned
cells of the query are assumed to be representative of those found in the reference.}
}
\value{
Tibble of query data with an added column of the predicted cell type, 'model_prediction'
}
\description{
Classify canonical cell types
}
