% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cyDefine.R
\name{cyDefine}
\alias{cyDefine}
\title{Run the full cyDefine pipeline}
\usage{
cyDefine(
  reference,
  query,
  markers,
  using_pbmc = FALSE,
  adapt_reference = ifelse(using_pbmc, TRUE, FALSE),
  min_f1 = 0.7,
  batch_correct = TRUE,
  ref.batch = NULL,
  xdim = 6,
  ydim = 6,
  identify_unassigned = TRUE,
  MAD_factor = 2.5,
  norm_method = "scale",
  covar = NULL,
  load_model = NULL,
  save_model = NULL,
  mtry = floor(length(markers)/3),
  splitrule = "gini",
  probability = TRUE,
  use.weights = TRUE,
  min.node.size = 1,
  num.trees = 600,
  num.threads = 1,
  mc.cores = NULL,
  save_adapted_reference = NULL,
  exclude_redundant = FALSE,
  unassigned_name = "unassigned",
  identify_type = "probability",
  probability_threshold = 0.5,
  train_on_unassigned = FALSE,
  seed = 332,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in
columns) if you want to apply a custom reference. If NULL, the Seurat
PBMC atlas will be adapted and used as reference.}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers present in both query
and reference.}

\item{using_pbmc}{Boolean indicating whether the Seurat PBMC atlas is
being supplied as reference. This info is needed to enable automated
screening of common marker names in \code{map_marker_names()} and tree-based
merging of cell type labels.}

\item{adapt_reference}{Boolean indicating whether the provided reference
should be adapted to match the query marker panel. Default: FALSE if
\code{using_pbmc} is FALSE, otherwise TRUE.}

\item{min_f1}{Minimum F1 score required for two cell types to be separated}

\item{batch_correct}{Boolean indicating whether or not you want to perform
batch correction via cyCombine}

\item{ref.batch}{Optional: Reference batch for ComBat.}

\item{xdim}{Integer or vector of integers, the x-dimension(s) of the SOM grid
(for "kohonen" and "flowsom"). If a vector, the workflow iterates.}

\item{ydim}{Integer or vector of integers, the y-dimension(s) of the SOM grid.
If a vector, the workflow iterates. Length should match \code{xdim} or be 1.}

\item{identify_unassigned}{Boolean indicating whether or not you want to
identify unassigned cells after classifying the cells.}

\item{MAD_factor}{Median average distance threshold of the outlier to its assigned population}

\item{norm_method}{Normalization method for cyCombine batch correction.
Should be either 'rank', 'scale', or 'qnorm'. Default is 'scale'.}

\item{covar}{Optional: Covariate column name or vector for correction.}

\item{load_model}{Optional: Path to an rda file of a previously trained model}

\item{save_model}{Optional: Path to save an rda file of the trained model}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{splitrule}{Splitting rule. For classification and probability estimation "gini", "extratrees" or "hellinger" with default "gini".
For regression "variance", "extratrees", "maxstat", "beta" or "poisson" with default "variance".
For survival "logrank", "extratrees", "C" or "maxstat" with default "logrank".}

\item{probability}{Grow a probability forest as in Malley et al. (2012).}

\item{use.weights}{(Default: FALSE) Boolean whether to use celltype frequency as class weights}

\item{min.node.size}{Minimal node size to split at. Default 1 for classification, 5 for regression, 3 for survival, and 10 for probability. For classification, this can be a vector of class-specific values.}

\item{num.trees}{Number of trees.}

\item{num.threads}{Number of threads. Use 0 for all available cores. Default is 2 if not set by options/environment variables (see below).}

\item{mc.cores}{Number of cores for parallel processing. Defaults to 1.}

\item{save_adapted_reference}{Default: NULL. Save the adapted reference in a
specified object name or as "adapted_reference.rds" in a specified path.
Usage: FAlSE/NULL - not stored.
TRUE/path - stored as "adapted_reference.rds" at "./" or path.
path/to/filename.rds - stored as "filename.rds" at path/to/.}

\item{exclude_redundant}{Exclude celltypes in the reference that are not
present in the query.}

\item{unassigned_name}{Name used for unassigned cells}

\item{identify_type}{(Default: probability) Type of outlier detection.
Options are (multiple allowed in conjunction):
"probability" - Cut cells with a lower probability than the \code{probability_threshold}.
"trained" - Train on unassigned in the reference.
"mad" - MAD threshold on mahalanobis distance in PC space per cell type.
"maha" - Same as "mad", but in marker space.
"sd" - Unassign cells with more than 20\% of markers that lay more than two standard deviations from reference expression}

\item{probability_threshold}{Threshold for random forest probability threshold. Use in conjunction with \code{unassigned_type = "probability"}.}

\item{train_on_unassigned}{Boolean indicating whether unassigned cells should be included in model training.
Recommended when reference and query are samples stemming from the same experiment and unassigned
cells of the query are assumed to be representative of those found in the reference.}

\item{seed}{desc}

\item{verbose}{Verbosity}

\item{...}{Additional optional parameters for batch correction.
See \code{?cyCombine::batch_correct} for options.}
}
\value{
Tibble of query data with added columns: "model_prediction"
indicating the canonical cell type predicted by the model, and
"predicted_celltype" indicating the final predicted cell type after
identifying unassigned cells, i.e. will either be "unassigned" or the same
as "model_prediction".
}
\description{
Run the full cyDefine pipeline
}
