% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cyDefine.R
\name{cyDefine}
\alias{cyDefine}
\title{Run the full cyDefine pipeline}
\usage{
cyDefine(
  reference = NULL,
  query,
  markers,
  adapt_PBMC_reference = TRUE,
  batch_correct = TRUE,
  identify_unassigned = TRUE,
  norm_method = "scale",
  covar = NULL,
  load_model = NULL,
  save_model = NULL,
  unassigned_name = "unassigned",
  train_on_unassigned = TRUE,
  n_cores = 2,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{adapt_PBMC_reference}{Boolean indicating whether or not you want to use the
Seurat PBMC atlas as reference, rather than supplying your own reference.}

\item{batch_correct}{Boolean indicating whether or not you want to perform
batch correction via cyCombine}

\item{identify_unassigned}{Boolean indicating whether or not you want to
identify unassigned cells after classifying the cells.}

\item{norm_method}{Normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'scale'}

\item{covar}{The covariate ComBat should use. Can be a vector or a column name in the input tibble.
If NULL, no covar will be used}

\item{load_model}{Optional: Path to an RDS file of a previously trained model}

\item{save_model}{Optional: Path to save an RDS file of the trained model}

\item{unassigned_name}{Name used for unassigned cells}

\item{train_on_unassigned}{Boolean indicating whether unassigned cells should be included in model training.
Recommended when reference and query are samples stemming from the same experiment and unassigned
cells of the query are assumed to be representative of those found in the reference.}

\item{n_cores}{The number of cores to use for parallel execution. Obs:
parallel execution only performed if the 'doParallel' package is installed!}

\item{seed}{The seed to use when creating the SOM.}

\item{verbose}{Verbosity}
}
\value{
Tibble of query data with added columns: "model_prediction"
indicating the canonical cell type predicted by the model, and
"predicted_celltype" indicating the final predicted cell type after
identifying unassigned cells, i.e. will either be "unassigned" or the same
as "model_prediction".
}
\description{
Run the full cyDefine pipeline
}
