% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{get_merged_label}
\alias{get_merged_label}
\title{Get label of a merged population}
\usage{
get_merged_label(populations)
}
\arguments{
\item{populations}{Character vector of individual cell types in merged population}
}
\value{
Label of merged population
}
\description{
Get label of a merged population
}
\seealso{
Other adapt: 
\code{\link{map_marker_names}()},
\code{\link{merge_populations}()}
}
\concept{adapt}
