% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{identify_similar_populations}
\alias{identify_similar_populations}
\title{Identify indistinguishable cell type populations in the reference, based on
query marker panel}
\usage{
identify_similar_populations(
  reference,
  markers,
  num.threads = 4,
  mtry = 4,
  min_f1 = 0.7,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{desc}

\item{markers}{Character vector of available markers present in both query
and reference.}

\item{num.threads}{Number of threads. Use 0 for all available cores. Default is 2 if not set by options/environment variables (see below).}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{min_f1}{desc}

\item{seed}{desc}

\item{verbose}{desc}
}
\value{
A tibble of similar populations
}
\description{
Identify indistinguishable cell type populations in the reference, based on
query marker panel
}
\note{
param population_pairs Data frame of two character columns with one cell
type in each, and rows representing all pairs of cell types to be tested.
}
