% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{identify_similar_populations}
\alias{identify_similar_populations}
\title{Identify indistinguishable cell type populations based on query markers}
\usage{
identify_similar_populations(
  population_pairs,
  reference,
  min_f1 = 0.6,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Identify indistinguishable cell type populations based on query markers
}
