% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_identify_unassigned.R
\name{identify_unassigned}
\alias{identify_unassigned}
\title{Identify unassigned cells}
\usage{
identify_unassigned(
  reference,
  query,
  markers,
  train_on_unassigned = TRUE,
  unassigned_name = "unassigned",
  pct_expl_var = 0.95,
  MAD_factor = 2.5,
  seed = 332,
  verbose = TRUE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Already classified... ???}

\item{markers}{Character vector of available markers}

\item{train_on_unassigned}{Boolean indicating whether unassigned cells should be included in model training.
Recommended when reference and query are samples stemming from the same experiment and unassigned
cells of the query are assumed to be representative of those found in the reference.}

\item{unassigned_name}{Name used for unassigned cells}

\item{pct_expl_var}{???}

\item{MAD_factor}{???}

\item{seed}{Random seed}

\item{verbose}{Verbosity}

\item{tune_params_on_unassigned}{???}
}
\description{
Identify unassigned cells
}
