% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{map_marker_names}
\alias{map_marker_names}
\title{Map query marker names to reference marker names}
\usage{
map_marker_names(
  use_seurat = TRUE,
  ref_markers = NULL,
  query,
  query_markers,
  map_specific_from = NULL,
  map_specific_to = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ref_markers}{Character vector of reference markers}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{query_markers}{Character vector of query markers}

\item{map_specific_from}{Optional: used together with map_specific_to. Character vector of specific marker(s) of the query to manually map to specific marker(s) of the reference (given by map_specific_to)}

\item{map_specific_to}{Optional: used together with map_specific_from. Markers of the reference that the markers given by map_specific_from should be mapped to}

\item{verbose}{Verbosity}
}
\value{
Tibble of query data modified to make marker names match reference markers
}
\description{
Map query marker names to reference marker names
}
\seealso{
Other adapt: 
\code{\link{get_merged_label}()},
\code{\link{merge_populations}()}
}
\concept{adapt}
