% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_adapt_reference.R
\name{merge_populations}
\alias{merge_populations}
\title{Merge groups of similar cell populations}
\usage{
merge_populations(populations_to_merge, reference, using_pbmc = TRUE)
}
\arguments{
\item{populations_to_merge}{A data frame of clustered populations}

\item{reference}{Data frame of reference data (cells in rows, markers in columns)}

\item{using_pbmc}{Boolean indicating whether the Seurat PBMC atlas is
being supplied as reference. This info is needed to enable automated
screening of common marker names in \code{map_marker_names()} and tree-based
merging of cell type labels.}

\item{name_unassigned_if_merged}{Cell types of the labeling tree, where if merged, the merged population should rather be named 'unassigned'}
}
\description{
Merge groups of similar cell populations
}
\seealso{
Other adapt: 
\code{\link{get_merged_label}()},
\code{\link{map_marker_names}()}
}
\concept{adapt}
