% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Make bar plot of predicted cell type abundances}
\usage{
plot_abundance(
  predicted_populations,
  colors = NULL,
  return_data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{predicted_populations}{Character vector of (predicted) populations to plot abundance for}

\item{colors}{Optional: Named list of colors per cell type. Values should be
HEX colors and names should be the unique cell types}

\item{return_data}{Boolean indicating if the data used for plotting should be
returned rather than a ggplot2 object}

\item{verbose}{Verbosity}
}
\description{
Make bar plot of predicted cell type abundances
}
