% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{plot_umap}
\alias{plot_umap}
\title{Visualize reference and query by UMAPs}
\usage{
plot_umap(
  reference,
  query = NULL,
  markers = NULL,
  colors = NULL,
  query_color_col = "predicted_celltype",
  build_umap_on = c("both", "reference"),
  metric = "euclidean",
  shuffle = TRUE,
  down_sample = TRUE,
  sample_n = 10000,
  return_data = FALSE,
  seed = 332,
  verbose = TRUE,
  title = c("Reference", "Query - predicted cell types"),
  col = c("celltype", "batch"),
  add_centroids = c(FALSE, "text", "label"),
  highlight_labels = FALSE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{colors}{Optional: Named list of colors per cell type. Values should be
HEX colors and names should be the unique cell types}

\item{query_color_col}{Column to color by. Default "predicted_celltype"}

\item{build_umap_on}{Character describing what data should be applied for
performing the UMAP embedding. Options are "both" (default), "reference",
and "query". "Both" implies that the UMAP embedding is done on both the
reference and query data.}

\item{shuffle}{Boolean indicating whether rows should be shuffled before plotting}

\item{down_sample}{Boolean indicating whether cells should be down-sampled}

\item{sample_n}{Number of cells to sample from each of reference and query,
if down_sample = TRUE}

\item{return_data}{Boolean indicating if the data used for plotting should be
returned rather than a ggplot2 object}

\item{seed}{Random seed}

\item{verbose}{Verbosity}
}
\value{
ggplot2 plot
}
\description{
Visualize reference and query by UMAPs
}
