% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_visualize.R
\name{plot_umap}
\alias{plot_umap}
\title{Visualize reference and query by UMAPs}
\usage{
plot_umap(
  reference,
  query = NULL,
  markers = NULL,
  colors = NULL,
  col = NULL,
  query_col = "predicted_celltype",
  ref_col = "celltype",
  build_umap_on = c("both", "reference"),
  metric = "euclidean",
  shuffle = TRUE,
  down_sample = TRUE,
  sample_n = 10000,
  return_data = FALSE,
  seed = 332,
  verbose = TRUE,
  title = c("Reference", "Query - predicted cell types"),
  legend_title = "Cell type",
  add_centroids = c(FALSE, TRUE, "text", "label"),
  highlight_labels = FALSE
)
}
\arguments{
\item{reference}{Tibble of reference data (cells in rows, markers in columns)}

\item{query}{Tibble of query data (cells in rows, markers in columns)}

\item{markers}{Character vector of available markers}

\item{colors}{Optional: Named list of colors per cell type. Values should be
HEX colors and names should be the unique cell types}

\item{col}{Column to color by.}

\item{query_col}{Column to color query by. Default "predicted_celltype"}

\item{ref_col}{Column to color reference by. Default "celltype"}

\item{build_umap_on}{Character describing what data should be applied for
performing the UMAP embedding. Options are "both" (default), "reference",
and "query". "Both" implies that the UMAP embedding is done on both the
reference and query data.}

\item{shuffle}{Boolean indicating whether rows should be shuffled before plotting}

\item{down_sample}{Boolean indicating whether cells should be down-sampled}

\item{sample_n}{Number of cells to sample from each of reference and query,
if down_sample = TRUE}

\item{return_data}{Boolean indicating if the data used for plotting should be
returned rather than a ggplot2 object}

\item{seed}{Random seed}

\item{verbose}{Verbosity}

\item{title}{Character string for plot title}

\item{legend_title}{Manually assigned legend title.}

\item{add_centroids}{Character or logical indicating whether to add centroids.
Options are "text"/TRUE, "label", or FALSE (default)}

\item{highlight_labels}{Logical indicating whether to highlight centroid labels
with ellipses (default FALSE)}
}
\value{
ggplot2 plot
}
\description{
Visualize reference and query by UMAPs
}
